pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "investor-relation-automation"
version = "0.1.0"
description = "Python-based agentic system that monitors Indonesian Stock Exchange (IDX) announcements, parses PDFs, and provides AI-powered financial sentiment analysis using OpenRouter LLMs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright==1.41.2",
    "PyMuPDF==1.23.8",
    "requests==2.31.0",
    "APScheduler==3.10.4",
    "python-dotenv==1.0.1",
    "httpx==0.27.0",
    "pymongo[srv]>=4.17.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "playwright-stealth>=2.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "config"]
